requirements.txt
flask
flask-cors
groq
pillow
pytesseract
python-dotenv
tesseract